requirements.txt
python-dotenv
openai
agent-framework --pre